# improve performance of `isapprox(::AbstractArray, ::AbstractArray)` when `rtol = 0`