The default value of `rtol` for `isapprox` depends on whether `atol` is zero or not; if `atol` is nonzero, `rtol` is zero (#22742).

The "normal-path" `isapprox(x, y; atol, rtol` check is just `norm(x-y) <= max(atol, rtol*max(norm(x), norm(y)))`. If `rtol = 0`, however, we are then redundantly computing the norm of both `x` and `y`, which can be expensive for arrays (unlike numbers).

This change simply puts a branch in to check `iszero(rtol)`: if so, it avoids computing the norms redundantly. This speeds up `isapprox(X, Y; atol ≠ 0)` considerably:

```jl
julia> A = rand(3, 3)
julia> @btime isapprox($A, $A; atol=1e-11);
  124.592 ns (1 allocation: 128 bytes) # master
julia> @btime isapprox($A, $A; atol=1e-11);
  68.821 ns (1 allocation: 128 bytes) # PR

julia> B = @SMatrix rand(3, 3)
julia> @btime isapprox($B, $B; atol=1e-11);
  14.544 ns (0 allocations: 0 bytes) # master
julia> @btime isapprox($B, $B; atol=1e-11);
  6.900 ns (0 allocations: 0 bytes) # PR
```

The cost of the extra branch is not nothing, of course, but should be negligible in the context of other costs for most array sizes.

---
There's also a small change to the `atol` default value: I noticed that care was being taken to initialize `rtol` to a type consistent with `x` and `y`; I figured the same might as well be done for `atol`.