@thchr 2022-11-10
Okay, that should solve my worry about dynamic dispatch, thanks @LilithHafner.

I think it's fine to keep the more refined `atol` default though, both for style consistency with efforts for `rtol` and to make it likely to do same-type `<=` comparisons.