@LilithHafner 2022-11-10
I'd still like to see an example where this increased complexity for `atol` is demonstrably better than the simpler `atol::Real=0` because there is a loss of readability changing `0` to `zero(promote_type(real(promote_leaf_eltypes(x)),real(promote_leaf_eltypes(y))))`.