@thchr 2022-11-10
Okay, we can keep the original `atol`: I don't feel strongly either way.

The current asymmetry between the care taken with the types of `rtol` and `atol` is odd though: in one case, we try quite hard, in the other, not at all.
Incidentally, the explicit `::Real` type-constraints on `rtol` and `atol` seem unfortunate: e.g., this - and the `atol = 0` default makes it so that dimensionful numbers (e.g., `Furlong` or Unitful.jl values) cannot fall back to Base's `isapprox` methods. 
In fact, Unitful.jl seems to currently [copy-extend this entire method](https://github.com/PainterQubits/Unitful.jl/blob/ebad4497360693722c9dae0a6d6f28b5647ad4a6/src/quantities.jl#L294); I assume partly for these reasons (e.g., because a unitful number cannot compare to a non-unitful one; and also because an `AbstractQuantity` isn't a subtype of `Real`).